Finish 'preload_scripts' for connection arguments
The new configuration 'preload_scritps' allows users to load SQL scripts
from remote locations so that the connected database will be prepared
before executing any statements.This feature enable us to use DuckDB in-memory
as a local query engine in any BI systems like Superset and Redash.

The supported remote protocols are http://, https://, s3://, s3a://.
In addition to the remote locations,the feature allow us to provide parameters
to substitute the fetched preload scripts in case that we have some secret or custom
configurations which should not be kept in the remote files.

Example of configuration to use preload scripts feature:

```python
engine = create_engine(
    "duckdb:///:memory:",
    connect_args={
        "preload_scripts": {
            "location": "https://duckdb-preloader-example.com/mydataset-preploader.sql",
            "parameters": {
                "mysecret": "mysupersecret",
            },
        },
    },
)
```